+++ NEWS.md
@@ -1,3 +1,4 @@
+- Added `ypr_length_at_age()` to get lengths at specific ages for a given population.
 - Improved handling of unsustainable populations (no longer negative fish numbers).
 - Added `expand = TRUE` argument to `ypr_populations()` to allow control over all combinations.
 - Added parameters `k2`, `Linf2` and `L2` to allow for biphasic growth.